♻️ refactor: post submit 관련 함수 util로 분리
